added readme file parsing
These changes co-exist with the present behavior by only intervening if
the changelog file is missing. When the changelog is missing, this now
presumes the readme file should be used as a description. Each h2 (##)
section of the readme is used to populate a tab in the plugin-installer
view details. To include the changelog, just include a section called
“Changelog” in the readme. This allows for much more informative
descriptions for users of GitHub/Bitbucket plugins.